Big Endian mode:   compare_be.go 	Little Endian mode:   compare_le.go 	Move the Compare function to BE mode or LE mode:   ulid.go
